requirements.txt
requests
python-dotenv
setuptools
twine
wheel